pyproject.toml
[project]
name = "doc"
version = "0.2.0"
description = "HTML to Word converter with perfect style preservation"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-docx>=1.1.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
    "fastapi>=0.104.0",
    "uvicorn[standard]>=0.24.0",
]
